TMS: added cached variant of tree (includes patches)
* added plugins and tree to installable cache components (patch)
* cache import_id in ilObjectDataCache (patch)
* added lang vars for new cache components
* cache getChilds
* cache getChildsByType
* cache getChildsByTypeFilter
* cache getSubTreeIds
* cache getSubTree
* cache getSubtreeTypes
* cache getPathFull
* cache getPathId
* cache getNodeData
* cache isInTree
* cache getParentNodeData
* cache isGrandChild
* cache getParentId
* cache checkForParentType

ATTENTION: Do not cached `getChilds` and `getChildsByType` with ordering
or sorting parameters.

The method shows a super weird behaviour in the case where order and
direction parameters are used. It doesn't seem to actually order stuff
according to the parameter. The both parameters are rarely used anyway,
though. So I decided to implement a fallback to the original tree
implementation if the params are used.

ATTENTION: getSubTreeIDs creates another order when cached.

getSubTreeIds actually creates another order then the original implementation,
but the order seems to be uninteresting anyways.

I had to delete tree and object_reference-entries on the database for the
node_ids 14, 24 and 25. These seem to be some left-overs from a faulty
delete process in the database template.

ATTENTION: only works with ilMaterializedPathTree

ilNestedSetTree internally rearranges lft and rgt in table, which can't
be noticed by the cache. Nested set is outdated anyway...